server/requirements.txt
asyncpg
sqlalchemy[asyncio]
python-dotenv
sqlalchemy-utils
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]